Examples and some bugfixs
Only example 11 was not updated (for now) because the extension system is not the same.
