+++ examples/io/viewers/viewer3D-5-custom.cpp
@@ -44,7 +44,7 @@
 #include "DGtal/io/Color.h"
 #include "DGtal/io/DrawWithDisplay3DModifier.h"
 #include "DGtal/shapes/Shapes.h"
-#include "DGtal/io/viewers/Viewer3D.h"
+#include "DGtal/io/viewers/PolyscopeViewer3D.h"
 
 ///////////////////////////////////////////////////////////////////////////////
 
@@ -58,13 +58,8 @@ using namespace Z3i;
 
 int main( int argc, char** argv )
 {
-
- QApplication application(argc,argv);
-
- typedef Viewer3D<> MyViewer;
+ typedef PolyscopeViewer3D<> MyViewer;
  MyViewer viewer;
- viewer.show();
-
 
  Point p1( -1, -1, -2 );
  Point p2( 2, 2, 3 );
@@ -91,8 +86,8 @@ int main( int argc, char** argv )
   viewer << SetMode3D(domain.className(), "Paving");
   viewer << domain << MyViewer::updateDisplay;
 
-
- return application.exec();
+  viewer.show();
+  return 0; 
 }
 //                                                                           //
 ///////////////////////////////////////////////////////////////////////////////
