+++ examples/io/viewers/viewer3D-7-planes.cpp
@@ -50,7 +50,7 @@ $ ./examples/io/viewers/viewer3D-7-planes
 #include <iostream>
 #include "DGtal/base/Common.h"
 #include "DGtal/helpers/StdDefs.h"
-#include "DGtal/io/viewers/Viewer3D.h"
+#include "DGtal/io/viewers/PolyscopeViewer3D.h"
 #include "DGtal/geometry/surfaces/COBANaivePlaneComputer.h"
 ///////////////////////////////////////////////////////////////////////////////
 
@@ -75,8 +75,6 @@ displayPredicate( Viewer3D & viewer,
 int main( int argc, char** argv )
 {
   using namespace Z3i;
-
-  QApplication application(argc,argv);
   trace.beginBlock ( "Testing class COBANaivePlaneComputer" );
 
   unsigned int nbok = 0;
@@ -142,9 +140,8 @@ int main( int argc, char** argv )
   trace.emphase() << ( nbok == nb ? "Passed." : "Error." ) << endl;
   trace.endBlock();
 
-  typedef Viewer3D<> MyViewer;
+  typedef PolyscopeViewer3D<> MyViewer;
   MyViewer viewer;
-  viewer.show();
   Color red( 255, 0, 0 );
   Color green( 0, 255, 0 );
   Color grey( 200, 200, 200 );
@@ -161,8 +158,8 @@ int main( int argc, char** argv )
 
   viewer << MyViewer::updateDisplay;
 
-
-  return application.exec();
+  viewer.show();
+  return 0; 
 }
 //                                                                           //
 ///////////////////////////////////////////////////////////////////////////////
